+++ src/base/tests/Vec_ut.cpp
@@ -100,7 +100,7 @@ void VecTest() {
 
     {
         str::Builder v(0, nullptr);
-        for (size_t i = 0; i < 32; i++) {
+        for (int i = 0; i < 32; i++) {
             utassert(len(v) == i * 6);
             v.Append("lambd");
             if (i % 2 == 0)
